Renamed a file
The callibration program isnt complete. What we have now is a program
that can callibrate from an image. Although this image needs to be the
same resolution as the video if the intent is to use it for video. It
outputs a csv file with magnification and pixel to real mm conversion
factor.